// frontend/src/api/generated/model/getUsersParams.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { GetUsersRole } from './getUsersRole';

export type GetUsersParams = {
role?: GetUsersRole;
active?: boolean;
q?: string;
page?: number;
size?: number;
sort?: string;
};