// frontend/src/api/generated/model/userResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UserResponseRole } from './userResponseRole';

export interface UserResponse {
  id?: number;
  email?: string;
  fullName?: string;
  role?: UserResponseRole;
  isActive?: boolean;
  createdAt?: string;
  updatedAt?: string;
}